refactor: move `utils.ts` to `utils/google.ts`
