+++ lib/utils/google.d.ts
