+++ lib/utils/google.js
@@ -47,4 +47,4 @@ async function getWebsiteRank(website, keyword, opts) {
     return undefined;
 }
 exports.getWebsiteRank = getWebsiteRank;
-//# sourceMappingURL=utils.js.map
+//# sourceMappingURL=google.js.map
