+++ lib/utils/index.d.ts
@@ -1,0 +1,1 @@
+export { getWebsiteRank, WebsiteRank } from "./google";
