+++ lib/utils/index.js
@@ -1,0 +1,6 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.getWebsiteRank = void 0;
+var google_1 = require("./google");
+Object.defineProperty(exports, "getWebsiteRank", { enumerable: true, get: function () { return google_1.getWebsiteRank; } });
+//# sourceMappingURL=index.js.map
